helpers.js splitted
