+++ lib/helpers/allowedTextElements.js
@@ -1,0 +1,2 @@
+const allowedTextElements = ['i', 'em', 'b', 'strong', 'bold', 'span', 'a', 'code', 'br', 'sup', 'sub'];
+module.exports = allowedTextElements;
