+++ lib/helpers/index.js
@@ -1,0 +1,3 @@
+const requireDirectory = require('require-directory');
+
+module.exports = requireDirectory(module);
