+++ lib/helpers/isExcludedNode.js
@@ -1,0 +1,9 @@
+//function return tru if node is excluded from searching translations
+module.exports = function isExcludedNode(node) {
+    if ((node.tag === 'i' && node.attributes.find(a => a.name === 'class' && a.value?.chars?.indexOf('auditboard-icons') !== -1)) ||
+        (node.tag === 'span' && node.attributes.find(a => a.name === 'class' && a.value?.chars?.indexOf('clearfix') !== -1))
+    ) {
+        return true;
+    }
+    return false;
+}
