+++ lib/helpers/isText.js
@@ -1,0 +1,25 @@
+const isExcludedNode = require('./isExcludedNode');
+const iterateChildren = require('./iterateChildren');
+const allowedTextElements = require('./allowedTextElements');
+
+module.exports = function isText(node) {
+    if (node.type === 'TextNode' || node.tag === 'br' || node.type === 'MustacheStatement') {
+        return true;
+    }
+
+    //allow only allowed text elements
+    if (!allowedTextElements.includes(node.tag) ||
+        (node.modifiers && node.modifiers.length)) {
+        return false;
+    }
+
+    //exclude special nodes
+    if (isExcludedNode(node)) {
+        return false;
+    }
+
+
+    let children = Array.from(iterateChildren(node));
+    let everyText = children.every(n => isText(n));
+    return everyText;
+}
