+++ package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "ember-template-recast": "^5.0.3",
     "glob": "^7.2.0",
+    "require-directory": "^2.1.1",
     "semver": "^7.3.5",
     "yargs": "^17.2.1"
   },
